[feat]: Moved app to new db models
