+++ src/edulytica_api/crud/base/factory.py
