+++ src/edulytica_api/routers/auth.py
@@ -5,16 +5,18 @@ from fastapi import Response
 from fastapi.security import OAuth2PasswordRequestForm
 from src.edulytica_api.crud.token_crud import TokenCrud
 from src.edulytica_api.crud.user_crud import UserCrud
+from src.edulytica_api.crud.user_role_crud import UserRoleCrud
 from src.edulytica_api.database import get_session
 import src.edulytica_api.schemas.auth as auth_schemas
 from sqlalchemy.ext.asyncio import AsyncSession
 from fastapi import Depends, HTTPException, status
 from src.edulytica_api.auth.helpers.validators import password_validate
-from src.edulytica_api.models.models import User
+from src.edulytica_api.models import User
 from src.edulytica_api.auth.auth_bearer import refresh_token_auth, access_token_auth
 from src.edulytica_api.auth.helpers.utils import create_access_token, create_refresh_token, verify_password, \
     get_hashed_password, get_expiry
 from src.edulytica_api.settings import REFRESH_TOKEN_EXPIRE_MINUTES
+from src.edulytica_api.utils.default_enums import UserRoleDefault
 from src.edulytica_api.utils.logger import api_logs
 
 
@@ -27,14 +29,14 @@ async def login(
     form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
     session: AsyncSession = Depends(get_session)
 ):
-    user = await UserCrud.get_filtered_by_params(session=session, username=form_data.username)
+    user = await UserCrud.get_filtered_by_params(session=session, login=form_data.username)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Incorrect email or password")
 
     user = user[0]
-    password_validate(form_data.password, user.password)
+    password_validate(form_data.password, user.password_hash)
 
     access_token = create_access_token(user.id)
     checker = uuid.uuid4()
@@ -58,13 +60,20 @@ async def register_user(
 ):
     from sqlalchemy import or_
 
-    filter_user = or_(User.username == user.username, User.email == user.email)
+    filter_user = or_(User.login == user.username, User.email == user.email)
     existing_user = await UserCrud.get_filtered(session=session, filter=filter_user)
     if existing_user:
         raise HTTPException(status_code=400, detail="Email already registered")
 
     encrypted_password = get_hashed_password(user.password)
-    await UserCrud.create(session=session, username=user.username, email=user.email, password=encrypted_password)
+    role = await UserRoleCrud.get_filtered_by_params(session=session, name=UserRoleDefault.USER)
+    await UserCrud.create(
+        session=session,
+        login=user.username,
+        email=user.email,
+        password_hash=encrypted_password,
+        role_id=role.id
+    )
 
     return {"message": "user created successfully"}
 
@@ -76,11 +85,11 @@ async def change_password(
     session: AsyncSession = Depends(get_session)
 ):
     user = auth_data['user']
-    if not verify_password(request.old_password, user.password):
+    if not verify_password(request.old_password, user.password_hash):
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid old password")
 
     encrypted_password = get_hashed_password(request.new_password)
-    user.password = encrypted_password
+    user.password_hash = encrypted_password
     session.add(user)
     await session.commit()
 
@@ -117,8 +126,7 @@ async def refresh_token(
         session=session,
         record_id=token.id,
         refresh_token=refresh_token,
-        checker=checker,
-        status=True
+        checker=checker
     )
 
     response.set_cookie(
