+++ src/edulytica_api/schemas/llm_schema.py
@@ -3,9 +3,6 @@ from typing import List
 
 from fastapi import UploadFile
 from pydantic import BaseModel, ConfigDict
-import datetime
-
-from src.edulytica_api.models.models import DocumentReport
 
 
 class UploadFileData(BaseModel):
