+++ src/edulytica_api/schemas/system_schemas.py
@@ -1,33 +1,147 @@
 from datetime import datetime
+from typing import Optional
 from pydantic import BaseModel, ConfigDict, UUID4
+
+
+class _UserRoleCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    name: str
+
+
+class _UserRoleUpdate(_UserRoleCreate):
+    id: UUID4
+
+
+class _UserRoleGet(_UserRoleUpdate):
+    pass
 
 
 class _UserCreate(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
-    username: str
+    login: str
     email: str
-    password: str
+    password_hash: str
+    name: Optional[str] = None
+    surname: Optional[str] = None
+    organization: Optional[str] = None
+    role_id: UUID4
 
 
 class _UserUpdate(_UserCreate):
     id: UUID4
+    is_active: bool
 
 
 class _UserGet(_UserUpdate):
-    disabled: bool
-
     created_at: datetime
     updated_at: datetime
+
+
+class _CheckCodeCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    code: str
+    user_id: UUID4
+
+
+class _CheckCodeUpdate(_CheckCodeCreate):
+    id: UUID4
+
+
+class _CheckCodeGet(_CheckCodeUpdate):
+    created_at: datetime
+
+
+class _TicketCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    prompt: Optional[str] = None
+    shared: Optional[bool] = False
+    user_id: UUID4
+    ticket_status_id: UUID4
+    event_id: Optional[UUID4] = None
+    custom_event_id: Optional[UUID4] = None
+    document_id: UUID4
+    document_summary_id: Optional[UUID4] = None
+    document_report_id: Optional[UUID4] = None
+
+
+class _TicketUpdate(_TicketCreate):
+    id: UUID4
+
+
+class _TicketGet(_TicketUpdate):
+    created_at: datetime
+
+
+class _TicketStatusCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    name: str
+
+
+class _TicketStatusUpdate(_TicketStatusCreate):
+    id: UUID4
+
+
+class _TicketStatusGet(_TicketStatusUpdate):
+    pass
+
+
+class _DocumentCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    id: Optional[UUID4]
+    file_path: str
+    user_id: UUID4
+
+
+class _DocumentUpdate(_DocumentCreate):
+    pass
+
+
+class _DocumentGet(_DocumentUpdate):
+    created_at: datetime
+
+
+class _DocumentSummaryCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    id: Optional[UUID4]
+    file_path: str
+
+
+class _DocumentSummaryUpdate(_DocumentSummaryCreate):
+    pass
+
+
+class _DocumentSummaryGet(_DocumentSummaryUpdate):
+    created_at: datetime
+
+
+class _DocumentReportCreate(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
+
+    id: Optional[UUID4]
+    file_path: str
+
+
+class _DocumentReportUpdate(_DocumentReportCreate):
+    pass
+
+
+class _DocumentReportGet(_DocumentReportUpdate):
+    created_at: datetime
 
 
 class _TokenCreate(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
-    user_id: UUID4
     refresh_token: str
     checker: UUID4
-    status: bool
+    user_id: UUID4
 
 
 class _TokenUpdate(_TokenCreate):
@@ -35,36 +149,35 @@ class _TokenUpdate(_TokenCreate):
 
 
 class _TokenGet(_TokenUpdate):
-    created_date: datetime
+    created_at: datetime
 
 
-class _ResultFilesCreate(BaseModel):
+class _EventCreate(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
-    file: str
-    user_id: UUID4
-    ticket_id: UUID4
+    name: str
+    description: Optional[str] = None
 
 
-class _ResultFilesUpdate(_ResultFilesCreate):
+class _EventUpdate(_EventCreate):
     id: UUID4
 
 
-class _ResultFilesGet(_ResultFilesUpdate):
-    data_create: datetime
+class _EventGet(_EventUpdate):
+    created_at: datetime
 
 
-class _TicketsCreate(BaseModel):
+class _CustomEventCreate(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
-    ticket_type: str
+    name: str
+    description: Optional[str] = None
     user_id: UUID4
-    status_id: int
 
 
-class _TicketsUpdate(_TicketsCreate):
+class _CustomEventUpdate(_CustomEventCreate):
     id: UUID4
 
 
-class _TicketsGet(_TicketsUpdate):
-    created_date: datetime
+class _CustomEventGet(_CustomEventUpdate):
+    created_at: datetime
